TF-1710 Move all method of PresentationMailbox object to extension to sync
