+++ lib/features/mailbox_dashboard/presentation/model/search/search_email_filter.dart
@@ -6,6 +6,7 @@ import 'package:jmap_dart_client/jmap/core/filter/operator/logic_filter_operator
 import 'package:jmap_dart_client/jmap/core/utc_date.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_filter_condition.dart';
 import 'package:model/extensions/email_filter_condition_extension.dart';
+import 'package:model/extensions/presentation_mailbox_extension.dart';
 import 'package:model/mailbox/presentation_mailbox.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/search/email_receive_time_type.dart';
 import 'package:tmail_ui_user/features/thread/domain/model/search_query.dart';
