+++ lib/features/mailbox_dashboard/presentation/widgets/email_quick_search_item_tile_widget.dart
@@ -5,6 +5,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
 import 'package:model/email/presentation_email.dart';
 import 'package:model/extensions/presentation_email_extension.dart';
+import 'package:model/extensions/presentation_mailbox_extension.dart';
 import 'package:model/mailbox/presentation_mailbox.dart';
 
 class EmailQuickSearchItemTileWidget extends StatelessWidget {
